Add a job that cleans up metron_agent's redirection
metron_agent will configure rsyslogd to forward logs via relp when
`syslog_daemon_config.enable` is `true`, but when it is `false` it
ignores the file, rather than cleans it up.  This means when we
reconfigure metron_agent to no longer handle rsyslog forwarding, it may
still be forwarding due to the remnant configuration file.

Here we add a complementary job that when `syslog_daemon_config.enable`
is `false` will remove the configuration and restart rsyslogd, if that
config file happens to be present.

We implement this cleanup in a monit-managed service, so we can use
monit's `depends on` to influnce inter-service dependencies where
necessary.